refactor: implement authentication, profile, and task management controllers and services
